Adding a new ASCII writer option LogAscii::size_limit_hint.
This is a a hint for a maximum size that a log entry's final textual
representation should not substantially exceed. It's not a hard limit,
but the writer will start truncating the output if it's reached. The
writer will however still produce syntactically correct output.

When truncating strings, a marker is inserted that says how much data
was skipped. Sets/vectors just stop logging elements when the limit is
reached, as there's no way to insert a marker without breaking the
format.

Implemented for both standard and JSON output.