Update POP3.cc
This pull request is similar to #79. On line 883 array 'pop3_cmd_word[1]' is accessed at index 19, which is out of bounds. Similarly, I believe this is a bug; however, I don't know how to correct it yet (and the ability to submit an Issue still appears to be disabled).

Found by https://github.com/bryongloden/cppcheck