Simplify how packets go into the SMB analyzer.
There was a copy of every single packet being made as it entered
the SMB analyzer.  This was because the entire analyzer used to be
handwritten (only using binpac for unit parsing, not for parsing
the whole protocol).  Now that the entire parser is binpac we don't
need that layer of indirection anymore and we're able to remove
a lot of complexity.