Fix a build failure on OpenBSD
The wordexp function doesn't exist in OpenBSD.  Skipping this
functionality only affects users who have bro installed in a directory
in the PATH and the directory name as it appears in PATH starts with
a tilde (e.g. "~/bin").  A simple workaround for affected users
would be to change the PATH environment variable to not contain any
tildes.