Experimental support for asynchronous function calls.
This adds a new "async" keyword that teaches Bro to suspend the
execution of asynchronous functions until the result is available.
This currently works for all the BiFs that used to require wrapping
into "when" statements:

    # cat a.bro
    event bro_init()
        {
        print (async lookup_hostname("www.icir.org"));
        print (async lookup_addr(131.159.14.1));
        }

    # bro ./a.bro
    {
    192.150.187.12
    }
    erdbeerschnitzel.net.in.tum.de

This is *very* exprimental right now, I'm actually not sure yet if
this is going to work this way. The main thing to understand is the
way it's implemented: it's using some very low-level magic for
switching stacks so that Bro can freeze execution inside the middle of
an event handler and resume there later. I'm borrowing code from
libtask and HILTI here and can't tell yet if (1) performance is good
enough to make this viable, and (2) it's sufficiently portable (tested
it on Linux and macOS so far, where it seems to work)

This comes with one small semantic change for "when" statements: if
there are "when" statements still pending at termination, Bro will
wait for them to complete before exiting.

TOOD:
     - Measure performance
     - Leak checking
     - "return when" is probably broken (we don't have a test for that)
     - At least the DNS functions occasionally neither return a result
       nor timeout it seems; that lets Bro hang.