Fix for successive invocations of Broker::enable
When Broker::enable(flags) is called multiple times from the scriptland,
only the first invocation will cause an endpoint to be created and the
handed over endpoint flags will be set. Every additional invocation
will be ignored. This is bad in cases that you want to enable
additional flags, e.g., auto_publish or auto_routing, in a script that
is initialized later on.
The fix now simply checks for each invocation of Broker::enable(flags)
if an endpoint was already created and if yes it will check if the
flags have changed. If the new flags are different from the old ones,

the new flags will be set.