Adding/extending retrying for Broker::{listen,peer}.
Broker::listen() now retries in regular intervals if it fails due to
the port is being open.

Broker::peer() gains back it 3rd parameter to specify an interval for
retrying if a connection cannot be established, or breaks (that was
hardcoded to 10s before).

Default for both is retrying every 30s.