Versioning Broker messages at the Bro-level.
Each message now comes with a Bro-level protocol version. This will
help if we need to change how Bro-level activity is represented. Doing
this on a per-message basis ensure that we catch differences even if
messages get routed througn other nodes. It's additional overhead, but
I don't think it will matter much. Doing this on