Strip off www for easier intel matching
This will ease the pain if people using the Intel Framework aren't using a strict formatting standard and will allow for matches of domain.com if they appear in the http.log as www.domain.com. 

Could probably be expanded out to include ftp, www2, www3, etc. if need be.