Fix Software_Version_Change notice message.
The "old" version displayed in this notice's message was actually just
the "new" version repeated.  This happened because the
"Software::tracked" table was updated before the log event is raised, so
there was actually no way to get the "old" version inside the log event
handler.

This also adds a new event, "Software::version_change", that can is
raised whenever a version change is detected.