Improve builtin_error() messages
For errors occurring in BIFs called from a wrapper function, these now
give the call location of the wrapper function instead of the BIF.

This relies on naming "internal" BIFs with a leading double-underscore
and then calling them from a single wrapper function of the same name
without the leading underscores.

e.g. Broker::__listen is called from Broker::listen, if there's an
error to report from the BIF, it would indicate the call site of
Broker::listen.

This change also starts tracking the entire script-land call stack
instead of just the last function call.