Fix Known scripts to be able to use alternate implemenation
And run the external test suite using the alternate implementation
due to data stores behaving differently when running on offline pcaps.
E.g. expirations are based on wall time, not packet time, and timeouts
(which *are* based on packet time) may occur when the store is still
initializing due to a large interval of packet time passing.